Taost message (#25)
-Added toast message by CodeSeven https://github.com/CodeSeven/toastr
-Added toast message for both succes and fails on the Orders and Home page
-Fixed some issue that i found while debugging (Sbaroi helped me)

+++ DeliveryCompany/Services/Services/OrdersService.cs
@@ -96,7 +96,10 @@ namespace Services.Services
             _repositoryWrapper.PackageRepository.Delete(package);
             _repositoryWrapper.Save();
             var delivery = _repositoryWrapper.DeliveryRepository.FindByCondition(d => d.OrderId.Equals(id)).FirstOrDefault();
-            _repositoryWrapper.DeliveryRepository.Delete(delivery);
+            if (delivery is not null)
+            {
+                _repositoryWrapper.DeliveryRepository.Delete(delivery);
+            }
             _repositoryWrapper.Save();
             return true;
         }
